+++ src/OpenBudget.Model.Tests/SerializationTests.cs
@@ -1,5 +1,5 @@
 ﻿using OpenBudget.Model.Entities;
-using OpenBudget.Model.Event;
+using OpenBudget.Model.Events;
 using OpenBudget.Model.EventStream;
 using OpenBudget.Model.Infrastructure;
 using NUnit.Framework;
