+++ src/OpenBudget.Model/BudgetStore/MemoryEventStore.cs
@@ -1,11 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using OpenBudget.Model.Event;
+﻿using OpenBudget.Model.Events;
 using OpenBudget.Model.EventStream;
+using OpenBudget.Model.Infrastructure;
+using System;
+using System.Collections.Generic;
 using System.Linq;
 
-namespace OpenBudget.Model.Infrastructure
+namespace OpenBudget.Model.BudgetStore
 {
     public class MemoryEventStore : IEventStore
     {
