+++ src/OpenBudget.Model/EventStream/FileEventStream.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
-using OpenBudget.Model.Event;
+using OpenBudget.Model.Events;
 using System.IO;
 
 namespace OpenBudget.Model.EventStream
