+++ src/OpenBudget.Model/EventStream/IEventStream.cs
@@ -1,4 +1,4 @@
-﻿using OpenBudget.Model.Event;
+﻿using OpenBudget.Model.Events;
 using OpenBudget.Model.Infrastructure;
 using System;
 using System.Collections.Generic;
