+++ src/OpenBudget.Model/EventStream/MemoryEventStream.cs
@@ -1,17 +1,17 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
-using OpenBudget.Model.Event;
+using OpenBudget.Model.Events;
 using OpenBudget.Model.EventStream;
 
 namespace OpenBudget.Model.Infrastructure
 {
-    public class InMemoryEventStream : IEventStream
+    public class MemoryEventStream : IEventStream
     {
         private Func<IEnumerable<ModelEvent>> _intializeEventIterator;
         private EventStreamHeader _header;
 
-        public InMemoryEventStream(Func<IEnumerable<ModelEvent>> intializeEventIterator, EventStreamHeader header)
+        public MemoryEventStream(Func<IEnumerable<ModelEvent>> intializeEventIterator, EventStreamHeader header)
         {
             _intializeEventIterator = intializeEventIterator;
             _header = header;
