+++ src/OpenBudget.Model/Infrastructure/Entities/NoCreateEntity.cs
@@ -1,4 +1,4 @@
-﻿using OpenBudget.Model.Event;
+﻿using OpenBudget.Model.Events;
 using System;
 using System.Collections.Generic;
 using System.Text;
