+++ src/OpenBudget.Model/Infrastructure/Entities/NoCreateEntityGenerator.cs
@@ -1,4 +1,4 @@
-﻿using OpenBudget.Model.Event;
+﻿using OpenBudget.Model.Events;
 using OpenBudget.Model.Util;
 using System;
 using System.Collections.Generic;
