--- src/OpenBudget.Model/Infrastructure/IHandler.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace OpenBudget.Model.Infrastructure
-{
-    public interface IHandler<TMessage>
-    {
-        void Handle(TMessage message);
-    }
-}
