+++ src/OpenBudget.Model/Infrastructure/Messaging/IHandler.cs
@@ -1,0 +1,7 @@
+﻿namespace OpenBudget.Model.Infrastructure.Messaging
+{
+    public interface IHandler<TMessage>
+    {
+        void Handle(TMessage message);
+    }
+}
