+++ src/OpenBudget.Model/Serialization/EventTypeBinder.cs
@@ -1,6 +1,6 @@
 ﻿using Newtonsoft.Json.Serialization;
 using OpenBudget.Model.Entities;
-using OpenBudget.Model.Event;
+using OpenBudget.Model.Events;
 using OpenBudget.Model.Infrastructure.Entities;
 using System;
 using System.Collections.Generic;
